requirements.txt
Flask==3.0.0
numpy==1.24.3
opencv-python==4.8.1.78
plotly==5.18.0
Pillow==10.1.0
